Forklift out the linkerd2 multicluster command
This is slightly disappointing, but probably not surprising: The
linkerd2 cli embeds YAML files, which must be codegenned *at build
time*, an operation which isn't supported for external go modules.

So: Here we add a shellscript and pre-build step to forklift out the
multicluster command definitions, patch up package references,
codegen *that*, and use that package.